requirements.txt
beautifulsoup4>=4.9.3
bs4>=0.0.1
motuclient>=1.8.8
netCDF4>=1.5.6
pandas>=1.2.3
python-dotenv>=0.15.0
PyYAML>=5.4.1
requests>=2.25.1
scipy>=1.6.1
siphon>=0.9
wget>=3.2
xarray==0.15.1
